scene12: raytrace transformed sphere
Raytracing in model space allows for any kind of transform to be applied
to the object. As long as the transform is reversible, the ray can be
transferred to model space. This allows specifying any oval or box with
two mat4 (or one, if inversed in shader).

Note: The transforms must be invertible. Creating 2D oval by scaling 3D
oval with one dimension of zero is non-invertible.